+++ python/fast_mlsirm/__init__.py
@@ -25,6 +25,7 @@ from .rt import fit_response_times as fit_response_times, RtFit as RtFit, fit_sp
 from .cdm import fit_cdm as fit_cdm, CdmFit as CdmFit, fit_gdina as fit_gdina, GdinaFit as GdinaFit, validate_q_matrix as validate_q_matrix, QMatrixValidation as QMatrixValidation, gdina_wald_selection as gdina_wald_selection, WaldModelSelection as WaldModelSelection, fit_ho_cdm as fit_ho_cdm, HoCdmFit as HoCdmFit
 from .mixture import fit_mixture as fit_mixture, MixtureFit as MixtureFit
 from .crm import fit_crm as fit_crm, CrmFit as CrmFit
+from .rsm import fit_rsm as fit_rsm, RsmFit as RsmFit
 from .mixed import fit_mixed_items as fit_mixed_items, MixedFormatFit as MixedFormatFit, MixedItemParameters as MixedItemParameters
 from .lltm import fit_lltm as fit_lltm, LltmFit as LltmFit
 from .testlet import fit_testlet as fit_testlet, TestletFit as TestletFit
@@ -107,6 +108,8 @@ __all__ = [
     "MixtureFit",
     "fit_crm",
     "CrmFit",
+    "fit_rsm",
+    "RsmFit",
     "fit_mixed_items",
     "MixedFormatFit",
     "MixedItemParameters",